--- lib/pages/blogs
@@ -1,1 +1,0 @@
-
