--- lib/pages/hosteldetails.dart
@@ -1,7 +1,0 @@
-class hostelDetails{
-   late String hostelName;
-   late String block;
-  late String floor;
-  late String roomNo;
-  hostelDetails(this.hostelName,this.block,this.floor,this.roomNo);
-}
