+++ lib/pages/login.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:onestop_dev/globals/myColors.dart';
-import 'package:onestop_dev/globals/myFonts.dart';
-import 'package:onestop_dev/globals/mySpaces.dart';
-import 'package:onestop_dev/globals/sizeConfig.dart';
+import 'package:onestop_dev/globals/my_colors.dart';
+import 'package:onestop_dev/globals/my_fonts.dart';
+import 'package:onestop_dev/globals/my_spaces.dart';
+import 'package:onestop_dev/globals/size_config.dart';
 import 'package:onestop_dev/stores/login_store.dart';
-import 'package:onestop_dev/widgets/powerup.dart';
+import 'package:onestop_dev/widgets/ui/powerup.dart';
 import 'package:provider/provider.dart';
