+++ lib/widgets/ui/appbar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:onestop_dev/globals/myColors.dart';
-import 'package:onestop_dev/globals/myFonts.dart';
+import 'package:onestop_dev/globals/my_colors.dart';
+import 'package:onestop_dev/globals/my_fonts.dart';
 import 'package:onestop_dev/pages/qr.dart';
 
 // TODO: Make profile picture clickable and redirect to QR
