+++ lib/widgets/ui/powerup.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:onestop_dev/globals/myColors.dart';
-import 'package:onestop_dev/globals/myFonts.dart';
-import 'package:onestop_dev/globals/mySpaces.dart';
-import 'package:onestop_dev/globals/sizeConfig.dart';
+import 'package:onestop_dev/globals/my_colors.dart';
+import 'package:onestop_dev/globals/my_fonts.dart';
+import 'package:onestop_dev/globals/my_spaces.dart';
+import 'package:onestop_dev/globals/size_config.dart';
 
 class PowerUp extends StatelessWidget {
   const PowerUp({Key? key}) : super(key: key);
